Use integration tests rather than a separate crate
