+++ Cargo.toml
@@ -125,7 +125,3 @@ inherits = "release"
 codegen-units = 1
 lto = true
 strip = "symbols"
-
-[workspace]
-members = [ ".", "./src/tests" ]
-default-members = [ "." ]
