+++ tests/milestones.rs
@@ -1,9 +1,11 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+mod common;
+
 use bee_block_stardust as bee;
 use chronicle::types::stardust::{block::payload::MilestoneId, util::payload::milestone::get_test_milestone_payload};
-use inx_chronicle_tests::connect_to_test_db;
+use common::connect_to_test_db;
 
 #[tokio::test]
 async fn test_milestones() {
