+++ tests/outputs.rs
@@ -1,6 +1,8 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+mod common;
+
 use bee_block_stardust as bee;
 use chronicle::{
     db::collections::{OutputMetadataResult, OutputWithMetadataResult},
@@ -16,7 +18,7 @@ use chronicle::{
         },
     },
 };
-use inx_chronicle_tests::connect_to_test_db;
+use common::connect_to_test_db;
 
 #[tokio::test]
 async fn test_outputs() {
